+++ src/domains/admin/admin-dashboard/admin-dashboard-rating-participation/api/react-query-api/admin-dashboard-rating-participation-query-keys.ts
@@ -1,0 +1,3 @@
+export const ADMIN_DASHBOARD_RATING_PARTICIPATION_QUERY_KEYS = {
+  RATING_PARTICIPATION: "admin-dashboard-rating-participation-rate",
+} as const;